# src/swagger/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "EDU Science API Document",
    "description": "https://n108.wke.csie.ncnu.edu.tw/",
    "version": "1.0.0"
  },
  "host": "localhost:8888",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "consumes": [],
  "produces": [],
  "paths": {
    "/Learn/subject": {
      "get": {
        "tags": [
          "Learn"
        ],
        "summary": "取得Learn中的科目",
        "description": "",
        "parameters": [
          {
            "name": "cid",
            "in": "query",
            "type": "integer"
          }
        ],
        "responses": {}
      }
    },
    "/Learn/titlecontent": {
      "get": {
        "tags": [
          "Learn"
        ],
        "summary": "取得Learn中的主題以及次主題",
        "description": "",
        "parameters": [
          {
            "name": "cid",
            "in": "query",
            "type": "integer"
          }
        ],
        "responses": {}
      }
    },
    "/Learn/content": {
      "get": {
        "tags": [
          "Learn"
        ],
        "summary": "取得主題底下的所有學習內容",
        "description": "",
        "parameters": [
          {
            "name": "cid",
            "in": "query",
            "type": "integer"
          }
        ],
        "responses": {}
      }
    },
    "/Learn/videoNum": {
      "get": {
        "tags": [
          "Learn"
        ],
        "summary": "取得學習內容影片數量",
        "description": "",
        "parameters": [
          {
            "name": "pcid",
            "in": "query",
            "type": "integer"
          }
        ],
        "responses": {}
      }
    },
    "/Learn/video": {
      "get": {
        "tags": [
          "Learn"
        ],
        "summary": "取得學習內容底下的所有影片內容",
        "description": "",
        "parameters": [
          {
            "name": "cid",
            "in": "query",
            "type": "integer"
          }
        ],
        "responses": {}
      }
    },
    "/Learn/{cid}": {
      "get": {
        "tags": [
          "Learn"
        ],
        "summary": "取得單個學習內容",
        "description": "",
        "parameters": [
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {}
      }
    },
...
    "/Note/{vid}": {
      "get": {
        "tags": [
          "Note"
        ],
        "summary": "取得影片筆記",
        "description": "",
        "parameters": [
          {
            "name": "vid",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {}
      },
      "post": {
        "tags": [
          "Note"
        ],
        "summary": "新增單個筆記",
        "description": "",
        "parameters": [
          {
            "name": "vid",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "currenttime": {
                  "type": "number"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/Note/note": {
      "delete": {
        "tags": [
          "Note"
        ],
        "summary": "刪除單個筆記",
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nid": {
                  "type": "integer"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/Note/title": {
      "put": {
        "tags": [
          "Note"
        ],
        "summary": "編輯筆記標題",
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nid": {
                  "type": "integer"
                },
                "title": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/Note/content": {
      "put": {
        "tags": [
          "Note"
        ],
        "summary": "編輯筆記內容",
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nid": {
                  "type": "integer"
                },
                "content": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
...
    "/Member/videohistory": {
      "post": {
        "tags": [
          "Member"
        ],
        "summary": "增加觀看紀錄、次數",
        "description": "",
        "parameters": [
          {
            "name": "vid",
            "in": "query",
            "type": "integer"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "CurrentTime": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
...
    "/Member/Notehistory": {
      "get": {
        "tags": [
          "Member"
        ],
        "summary": "取得 mid 底下所有筆記紀錄",
        "description": "",
        "parameters": [
          {
            "name": "vid",
            "in": "query",
            "type": "integer"
          }
        ],
        "responses": {}
      }
    },
...
  },
...
}